refactor(winaudio): extract per-arm helpers from run_com_loop (#37)
`run_com_loop` arms used to inline the master endpoint / session manager
None-checks and the OS-callback-vs-synthetic-emit branching. Each arm
ran 10-15 lines of boilerplate, pushing the file past the project's
500-LOC budget once W1's session cache lifecycle was added.

- Extract `handle_master_scalar`, `handle_master_mute`,
  `handle_refresh_master`, `handle_refresh_sessions`,
  `handle_enumerate_sessions` (in addition to the existing
  `handle_set_session_scalar` / `handle_toggle_session_mute`).
- Move all per-command handlers and `register_session_events_for_all`
  into a new `com_handlers` submodule. Split the for-PID logic into
  `register_session_events_for_one` and `purge_stale_session_regs`
  for clarity.
- `com_thread.rs` drops from 599 to 378 LOC (under the split @ 400
  threshold); the match block in `run_com_loop` is now 49 lines of
  clean 3-5-line arms.

Closes #37
