+++ src/drivers/winaudio/com_thread.rs
@@ -13,7 +13,7 @@ use std::time::{Duration, Instant};
 
 use tokio::sync::mpsc;
 use tokio::sync::Mutex;
-use tracing::{debug, error, info, trace, warn};
+use tracing::{debug, error, info, warn};
 
 use windows::Win32::Media::Audio::Endpoints::IAudioEndpointVolumeCallback;
 use windows::Win32::Media::Audio::{
@@ -22,15 +22,20 @@ use windows::Win32::Media::Audio::{
 use windows::Win32::System::Com::{CoInitializeEx, CoUninitialize, COINIT_APARTMENTTHREADED};
 
 use super::callback::EndpointVolumeCallback;
+use super::com_handlers::{
+    handle_enumerate_sessions, handle_master_mute, handle_master_scalar, handle_refresh_master,
+    handle_refresh_sessions, handle_set_session_scalar, handle_toggle_session_mute,
+    register_session_events_for_all,
+};
 use super::master::MasterEndpoint;
-use super::session::{set_session_volume, toggle_session_mute, SessionInfo, SessionManager};
-use super::session_events::{NewSessionCallback, SessionEventsCallback};
+use super::session::{SessionInfo, SessionManager};
+use super::session_events::NewSessionCallback;
 
 /// Max age for the cached session enumeration before a hot-path access
 /// triggers a fresh re-enumerate. `OnSessionCreated` /
 /// `OnStateChanged(Expired)` proactively invalidate the cache, so this
 /// is a safety net for transient sessions that slipped between events.
-const SESSION_CACHE_TTL: Duration = Duration::from_secs(2);
+pub(super) const SESSION_CACHE_TTL: Duration = Duration::from_secs(2);
 
 #[derive(Debug)]
 pub enum AudioCmd {
@@ -187,9 +192,9 @@ impl ComThreadHandle {
 /// Per-PID registration: we keep the `IAudioSessionControl2` so we can
 /// unregister at shutdown, and the `IAudioSessionEvents` interface so the
 /// audio engine has a live reference to call back into.
-struct SessionReg {
-    control: IAudioSessionControl2,
-    events: IAudioSessionEvents,
+pub(super) struct SessionReg {
+    pub(super) control: IAudioSessionControl2,
+    pub(super) events: IAudioSessionEvents,
 }
 
 fn run_com_loop(
@@ -294,45 +299,22 @@ fn run_com_loop(
     while let Some(cmd) = cmd_rx.blocking_recv() {
         match cmd {
             AudioCmd::SetMasterScalar(scalar) => {
-                if let Some(ep) = endpoint.as_ref() {
-                    if let Err(e) = ep.set_volume_scalar(scalar) {
-                        warn!("set_volume_scalar({}) failed: {}", scalar, e);
-                    } else if !master_callback_active {
-                        // Callback registration failed at init — fall back
-                        // to the legacy synthetic emit so the X-Touch stays
-                        // in sync. Normal path skips this (see #41).
-                        let mute = ep.get_mute().unwrap_or(false);
-                        let _ = event_tx.try_send(AudioEvent::MasterVolumeChanged { scalar, mute });
-                    }
-                }
+                handle_master_scalar(endpoint.as_ref(), scalar, master_callback_active, &event_tx);
             },
             AudioCmd::ToggleMasterMute => {
-                if let Some(ep) = endpoint.as_ref() {
-                    let cur = ep.get_mute().unwrap_or(false);
-                    if let Err(e) = ep.set_mute(!cur) {
-                        warn!("set_mute failed: {}", e);
-                    } else if !master_callback_active {
-                        // See SetMasterScalar — same fallback when the OS
-                        // callback isn't wired.
-                        let scalar = ep.get_volume_scalar().unwrap_or(0.0);
-                        let _ = event_tx
-                            .try_send(AudioEvent::MasterVolumeChanged { scalar, mute: !cur });
-                    }
-                }
+                handle_master_mute(endpoint.as_ref(), master_callback_active, &event_tx);
             },
             AudioCmd::RefreshMaster => {
-                if let Some(ep) = endpoint.as_ref() {
-                    if let (Ok(scalar), Ok(mute)) = (ep.get_volume_scalar(), ep.get_mute()) {
-                        let _ = event_tx.try_send(AudioEvent::MasterVolumeChanged { scalar, mute });
-                    }
-                }
+                handle_refresh_master(endpoint.as_ref(), &event_tx);
             },
             AudioCmd::RefreshSessions => {
-                if let Some(mgr) = session_mgr.as_ref() {
-                    let sessions =
-                        register_session_events_for_all(mgr, &event_tx, &cmd_tx, &mut session_regs);
-                    sessions_cache = Some((Instant::now(), sessions));
-                }
+                handle_refresh_sessions(
+                    session_mgr.as_ref(),
+                    &event_tx,
+                    &cmd_tx,
+                    &mut session_regs,
+                    &mut sessions_cache,
+                );
             },
             AudioCmd::SetSessionScalar {
                 process_name_lc,
@@ -353,19 +335,7 @@ fn run_com_loop(
                 }
             },
             AudioCmd::EnumerateSessions { reply } => {
-                let names = session_mgr
-                    .as_ref()
-                    .and_then(|m| m.enumerate().ok())
-                    .map(|sessions| {
-                        let names = sessions
-                            .iter()
-                            .map(|s| s.process_name.clone())
-                            .filter(|n| !n.is_empty())
-                            .collect::<Vec<_>>();
-                        sessions_cache = Some((Instant::now(), sessions));
-                        names
-                    })
-                    .unwrap_or_default();
+                let names = handle_enumerate_sessions(session_mgr.as_ref(), &mut sessions_cache);
                 let _ = reply.send(names);
             },
             AudioCmd::Shutdown => {
@@ -403,196 +373,4 @@ fn run_com_loop(
     drop(endpoint);
     drop(session_mgr);
     unsafe { CoUninitialize() };
-}
-
-/// Enumerate all sessions, register an `IAudioSessionEvents` callback on
-/// each new one, push a `SessionVolumeSnapshot` so the consumer sees
-/// every session's current state immediately, then emit a single
-/// `ActiveSessionsChanged` summary at the end. Re-running this is safe:
-/// already-registered sessions are skipped (looked up by PID), and PIDs
-/// that have disappeared since the previous pass are unregistered to
-/// prevent the `session_regs` map from growing unbounded.
-///
-/// Returns the freshly enumerated `SessionInfo`s so callers can populate
-/// the cache used by hot-path `Set*` / `Toggle*` arms (#35).
-fn register_session_events_for_all(
-    mgr: &SessionManager,
-    event_tx: &mpsc::Sender<AudioEvent>,
-    cmd_tx: &mpsc::Sender<AudioCmd>,
-    session_regs: &mut HashMap<u32, SessionReg>,
-) -> Vec<SessionInfo> {
-    let sessions = match mgr.enumerate() {
-        Ok(s) => s,
-        Err(e) => {
-            warn!("session enumerate failed: {}", e);
-            return Vec::new();
-        },
-    };
-
-    let mut active_names: Vec<String> = Vec::with_capacity(sessions.len());
-    let mut active_pids: std::collections::HashSet<u32> =
-        std::collections::HashSet::with_capacity(sessions.len());
-
-    for s in &sessions {
-        if s.process_name.is_empty() {
-            continue;
-        }
-        active_pids.insert(s.pid);
-        if !active_names.contains(&s.process_name) {
-            active_names.push(s.process_name.clone());
-        }
-
-        // Always push the current state — the consumer is idempotent and
-        // this is what catches sessions whose volume changed while we
-        // weren't subscribed (or before our callback was registered).
-        let scalar = unsafe { s.volume.GetMasterVolume().unwrap_or(0.0) };
-        let mute = unsafe { s.volume.GetMute().map(|b| b.as_bool()).unwrap_or(false) };
-        let _ = event_tx.try_send(AudioEvent::SessionVolumeSnapshot {
-            process_name_lc: s.process_name.clone(),
-            scalar,
-            mute,
-        });
-
-        // Skip if we already have an events callback registered for this PID.
-        if session_regs.contains_key(&s.pid) {
-            continue;
-        }
-
-        let cb_impl =
-            SessionEventsCallback::new(event_tx.clone(), cmd_tx.clone(), s.process_name.clone());
-        let events: IAudioSessionEvents = cb_impl.into();
-        // Cloning a COM interface bumps the refcount; the registration
-        // map then holds an independent reference to the same control
-        // object that we also keep in the cache.
-        match unsafe { s.control.RegisterAudioSessionNotification(&events) } {
-            Ok(()) => {
-                debug!(
-                    "RegisterAudioSessionNotification(pid={}, {}) succeeded",
-                    s.pid, s.process_name
-                );
-                session_regs.insert(
-                    s.pid,
-                    SessionReg {
-                        control: s.control.clone(),
-                        events,
-                    },
-                );
-            },
-            Err(e) => warn!(
-                "RegisterAudioSessionNotification(pid={}, {}) failed: {:?}",
-                s.pid, s.process_name, e
-            ),
-        }
-    }
-
-    // Unregister callbacks for PIDs that have disappeared since the last
-    // pass. Prevents the map from growing unboundedly across long-lived
-    // sessions where many short-lived audio sources come and go.
-    let stale_pids: Vec<u32> = session_regs
-        .keys()
-        .filter(|pid| !active_pids.contains(pid))
-        .copied()
-        .collect();
-    for pid in stale_pids {
-        if let Some(reg) = session_regs.remove(&pid) {
-            if let Err(e) = unsafe { reg.control.UnregisterAudioSessionNotification(&reg.events) } {
-                trace!(
-                    "UnregisterAudioSessionNotification(pid={}) on stale session failed: {:?}",
-                    pid,
-                    e
-                );
-            }
-        }
-    }
-
-    // Atomic "active set" marker — the consumer uses this to swap its
-    // active-session view in one go.
-    let _ = event_tx.try_send(AudioEvent::ActiveSessionsChanged {
-        names_lc: active_names,
-    });
-
-    sessions
-}
-
-/// Return a reference to a cached session matching `process_name_lc`,
-/// re-enumerating once if the cache is missing/expired or the name
-/// isn't found (handles the race where a transient session disappeared
-/// after the last refresh). The returned reference borrows from
-/// `sessions_cache`, which is updated in place on a refresh. Returns
-/// `None` if no matching session exists after the refresh.
-fn cached_session<'a>(
-    mgr: &SessionManager,
-    sessions_cache: &'a mut Option<(Instant, Vec<SessionInfo>)>,
-    process_name_lc: &str,
-) -> Option<&'a SessionInfo> {
-    let needs_refresh = match sessions_cache.as_ref() {
-        None => true,
-        Some((ts, sessions)) => {
-            ts.elapsed() > SESSION_CACHE_TTL
-                || super::mapping::find_session(sessions, process_name_lc).is_none()
-        },
-    };
-    if needs_refresh {
-        match mgr.enumerate() {
-            Ok(sessions) => {
-                *sessions_cache = Some((Instant::now(), sessions));
-            },
-            Err(e) => {
-                warn!("session enumerate failed (cache refresh): {}", e);
-                return None;
-            },
-        }
-    }
-    let sessions = &sessions_cache.as_ref()?.1;
-    super::mapping::find_session(sessions, process_name_lc)
-}
-
-/// Apply `scalar` to the cached session matching `process_name_lc`,
-/// refreshing the cache on miss/TTL expiry. See #35.
-fn handle_set_session_scalar(
-    mgr: &SessionManager,
-    sessions_cache: &mut Option<(Instant, Vec<SessionInfo>)>,
-    process_name_lc: &str,
-    scalar: f32,
-) {
-    let Some(s) = cached_session(mgr, sessions_cache, process_name_lc) else {
-        trace!(
-            "No active session for '{}'; ignoring volume command",
-            process_name_lc
-        );
-        return;
-    };
-    if let Err(e) = set_session_volume(s, scalar) {
-        warn!(
-            "SetSessionVolume({}, {}) failed: {}",
-            process_name_lc, scalar, e
-        );
-    }
-}
-
-/// Toggle mute on the cached session matching `process_name_lc` and
-/// emit the resulting state. Sessions have no callback equivalent to
-/// `IAudioEndpointVolumeCallback`, so the mute LED would never refresh
-/// unless we push a snapshot here ourselves. See #35.
-fn handle_toggle_session_mute(
-    mgr: &SessionManager,
-    sessions_cache: &mut Option<(Instant, Vec<SessionInfo>)>,
-    process_name_lc: &str,
-    event_tx: &mpsc::Sender<AudioEvent>,
-) {
-    let Some(s) = cached_session(mgr, sessions_cache, process_name_lc) else {
-        return;
-    };
-    match toggle_session_mute(s) {
-        Ok(()) => {
-            let scalar = unsafe { s.volume.GetMasterVolume().unwrap_or(0.0) };
-            let mute = unsafe { s.volume.GetMute().map(|b| b.as_bool()).unwrap_or(false) };
-            let _ = event_tx.try_send(AudioEvent::SessionVolumeSnapshot {
-                process_name_lc: process_name_lc.to_string(),
-                scalar,
-                mute,
-            });
-        },
-        Err(e) => warn!("ToggleSessionMute({}) failed: {}", process_name_lc, e),
-    }
 }
